[RUMF-913] do not collect <input type="radio"> and <input type="checkbox"> values
After asking myself if we should mask values for radios with
corresponding names, I came to the conclusion that the value of radios and
checkboxes shouldn't matter during replay, so removing it altogether
would be simpler.

This commit also removes the 'isChecked' property on inputs other than
radio or checkbox, making records a bit smaller.